+++ src/backend/resolvers/mutations/marketBasket/category/create.ts
@@ -3,7 +3,7 @@ const axios = require('axios')
 const createMarketBasketCategory = async (
   parent,
   { input },
-  context,
+  { pulseDevDb },
   info
 ) => {
   const vegaInput = {
@@ -11,11 +11,27 @@ const createMarketBasketCategory = async (
     characteristics: [],
   }
 
-  return await axios.post('market-basket-surveys-categories/', vegaInput)
+  const createdMarketBasketCategory = await axios.post('market-basket-surveys-categories/', vegaInput)
     .then(({ data }) => data)
     .catch((e) => {
       throw new Error(JSON.stringify(e.response.data))
     })
+
+  await pulseDevDb.collection('marketBaskets').updateOne(
+    { '_id': createdMarketBasketCategory.market_basket },
+    {
+      $push: {
+        'categories': {
+          '_id': createdMarketBasketCategory.id,
+          'name': createdMarketBasketCategory.name,
+          'prompt': createdMarketBasketCategory.prompt,
+          'characteristics': [],
+        },
+      },
+    }
+  )
+
+  return createdMarketBasketCategory
 }
 
 export default createMarketBasketCategory