+++ src/backend/resolvers/mutations/marketBasket/category/delete.ts
@@ -3,7 +3,7 @@ import axios from 'axios'
 const deleteMarketBasketCategory = async (
   parent,
   { input: { id } },
-  context,
+  { pulseDevDb },
   info
 ) => {
   const categoryToBeDeleted = await axios.get(`market-basket-surveys-categories/${id}/`)
@@ -13,6 +13,19 @@ const deleteMarketBasketCategory = async (
     .catch((e) => {
       throw new Error(JSON.stringify(e.response.data))
     })
+
+  await pulseDevDb.collection('marketBaskets').updateOne(
+    { '_id': categoryToBeDeleted.market_basket },
+    {
+      $pull: {
+        'categories': { '_id': id },
+      },
+    }
+  )
+
+  await pulseDevDb.collection('marketBasketsSurveyAnswers').deleteMany(
+    { 'category._id': id }
+  )
 
   return categoryToBeDeleted
 }