+++ src/backend/typeDefs/mutations/marketBasket/category/update.ts
@@ -3,9 +3,10 @@ import { gql } from 'apollo-server-express'
 const updateMarketBasketCategoryTypeDefs = gql`
   input UpdateMarketBasketCategoryInput {
     id: ID!
-    name: String!
-    category_type: String!
+    name: String
+    category_type: String
     prompt: String
+    characteristics: [ID!]
   }
 `