+++ src/frontend/api/mutations/marketBaskets.ts
@@ -29,7 +29,19 @@ export const DELETE_MARKET_BASKET = gql`
       updated_at
       products_regimens
       team_subscriptions
-      categories
+      categories {
+        id
+        name
+        prompt
+        category_type
+        _order
+        characteristics
+        characteristics_full {
+          id
+          name
+          description
+        }
+      }
     }
   }
 `
@@ -45,7 +57,19 @@ export const CREATE_MARKET_BASKET = gql`
       updated_at
       products_regimens
       team_subscriptions
-      categories
+      categories {
+        id
+        name
+        prompt
+        category_type
+        _order
+        characteristics
+        characteristics_full {
+          id
+          name
+          description
+        }
+      }
     }
   }
 `
@@ -61,7 +85,19 @@ export const UPDATE_MARKET_BASKET = gql`
       updated_at
       products_regimens
       team_subscriptions
-      categories
+      categories {
+        id
+        name
+        prompt
+        category_type
+        _order
+        characteristics
+        characteristics_full {
+          id
+          name
+          description
+        }
+      }
     }
   }
 `